Ignore Null and Unknown in storage size validator
Using Crusoe resources in a child module triggers a validation error
when passing the disk size as an input variable.

In this scenario, the validator is called both before the attribute is
set and after. Most string validators in Hashicorp's
terraform-plugin-framework-validators repo [1], for example, check for
Null or Unknown and return.

This patch changes the storage size validator to do the same.

[1] https://github.com/hashicorp/terraform-plugin-framework-validators