feat: capacity in data.ib_network
This commit adds the `capacities` field from the Crusoe API response
model for `ib_network`s to the `ib_network` data source. This is
important for Terraform modules to be able to filter available networks
from the data source response based on the `slice_type` and the
avalability in the `quantity` field.

Note: this data is already readily available in the public Crusoe JSON
API; this commit merely translates it into the provider for Terraform
users to benefit from it.

Signed-off-by: squat <lserven@gmail.com>